backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow-cpu==2.15.0
numpy==1.24.3
pillow==10.0.1
geopy==2.3.0
h5py==3.9.0